# Conflicts:
#	AdyenPSP/AdyenThreeDSecureUtils.swift
#	Client/KarhooConfig.swift
#	KarhooUISDK.podspec
#	KarhooUISDK.xcodeproj/project.pbxproj
#	KarhooUISDK/CommonUI/UITexts.swift
#	KarhooUISDK/Controllers/NavigationController.swift
#	KarhooUISDK/Extensions/UIView+Extensions.swift
#	KarhooUISDK/Screens/AddressScreen/AddressViewController.swift
#	KarhooUISDK/Screens/BookingScreen/BookingRouter.swift
#	KarhooUISDK/Screens/BookingScreen/KarhooBookingViewController.swift
#	KarhooUISDK/Screens/BookingScreen/Map/EmptyMapBookingStrategy.swift
#	KarhooUISDK/Screens/BookingScreen/Map/KarhooBookingMapPresenter.swift
#	KarhooUISDK/Screens/CheckoutScreen/Checkout/KarhooCheckoutViewController.swift
#	KarhooUISDK/Screens/CheckoutScreen/Components/CheckoutViews/KarhooCheckoutHeaderView.swift
#	KarhooUISDK/Screens/QuoteList/QuoteListFilters/QuoteListFiltersViewController.swift
#	KarhooUISDK/Screens/QuoteList/QuoteListSort/QuoteListSortViewController.swift
#	KarhooUISDK/Screens/QuoteList/QuoteListViewController.swift
#	KarhooUISDK/Screens/Rides/RideDetails/RideDetailsViewController.swift
#	KarhooUISDK/Screens/Rides/RidesList/RidesListPresenter.swift
#	KarhooUISDK/Screens/Rides/RidesViewController.swift
#	KarhooUISDK/Translations/es.lproj/Localizable.strings
#	KarhooUISDK/Translations/it.lproj/Localizable.strings
#	KarhooUISDK/Workers/Utils.swift
#	KarhooUISDKTests/TestCases/Screens/BookingScreen/Map/BookingMapPresenterSpec.swift
#	KarhooUISDKTests/TestCases/Screens/BookingScreen/QuoteCategoryBar/KarhooQuoteCategoryBarPresenterSpec.swift
#	Package.resolved
#	Package.swift
#	Podfile
#	Podfile.lock
#	README.md

+++ KarhooUISDK/Extensions/UIView+Extensions.swift
@@ -168,9 +168,9 @@ extension UIView {
     }
 
     /// Add shadow with given opacity (default = 0.5) and radius (default from UIKit = 3) and 0.5,0.5 offset.
-    func addShadow(_ opacity: Float = Float(UIConstants.Alpha.shadow), radius: CGFloat = 3) {
+    func addShadow(_ opacity: CGFloat = UIConstants.Alpha.shadow, radius: CGFloat = 3) {
         layer.shadowColor = UIColor.black.cgColor
-        layer.shadowOpacity = opacity
+        layer.shadowOpacity = Float(opacity)
         layer.shadowOffset = CGSize.init(width: 0.5, height: 0.5)
         layer.masksToBounds = false
         layer.shadowRadius = radius
